+++ src/main/java/com/vroommates/VroomMates/model/tripmodel/TripRepository.java
@@ -2,5 +2,5 @@ package com.vroommates.VroomMates.model.tripmodel;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TripRepository extends JpaRepository<Trip, Long> {
+public interface TripRepository extends JpaRepository<Trip, Integer> {
 }
